Support multiple kubeconfig files (#5)
Add support for passing multiple kubeconfig files either via the
`KUBECONFIG` environment variable or the `-k/--kubeconfig` CLI option.
This relies on a custom implementation of `clientcmd.ConfigAccess` as
the one used by `kubectl` does not unify CLI inputs and implicitly leaks
environment variables into other areas of the codebase. As a result, it
only supports multiple files via an environment variable.

During testing, I found that combining multiple files each with a
current context specified did not cause all current contexts to be wiped
when considered individually. This is only a problem if the order of the
files changes as the first file(s) takes precedence when defining the
current context.

Other changes:
- Upgrade golangci-lint to v2.
- Add an `all` just recipe to simplify development.
- Add groups to just recipes to improve developer experience.
- Summarise unit test coverage and generate a HTML report.
- Add placeholder `Help` methods to subcommands.
- Add docs to `version` and `logging` packages.
- Add unit tests to `version` package.